requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-cloud-vision
PyMuPDF
opencv-python
numpy
httpx